SplitViewController added
SplitViewController functionality added in iPad app. its a universal
app now containing 2 different storyboards for iPad and iPhone. iPad
using splitViewController and in landscape it shows 2 panels left and
right while in portrait it hides left one and shows a button on the top
bar of the details view to show a popover containing left MVC